requirements.txt
discord.py==1.5.1
SQLALchemy==1.3.18